rename spi package
