+++ notifications/core-spi/src/main/kotlin/org/opensearch/notifications/spi/NotificationCoreExtension.kt
@@ -9,7 +9,7 @@
  * GitHub history for details.
  */
 
-package org.opensearch.notifications.corespi
+package org.opensearch.notifications.spi
 
 /**
  * SPI of Notification Core
