+++ notifications/core-spi/src/main/kotlin/org/opensearch/notifications/spi/model/SecureDestinationSettings.kt
@@ -9,10 +9,8 @@
  *  GitHub history for details.
  */
 
-package org.opensearch.notifications.corespi.model.destination
-/**
- * Supported notification destinations
- */
-enum class DestinationType {
-    CHIME, SLACK, CUSTOM_WEBHOOK, SMTP, SES, SNS
-}
+package org.opensearch.notifications.spi.model
+
+import org.opensearch.common.settings.SecureString
+
+data class SecureDestinationSettings(val emailUsername: SecureString, val emailPassword: SecureString)
