+++ notifications/core-spi/src/main/kotlin/org/opensearch/notifications/spi/model/destination/CustomWebhookDestination.kt
@@ -9,9 +9,9 @@
  *  GitHub history for details.
  */
 
-package org.opensearch.notifications.corespi.model.destination
+package org.opensearch.notifications.spi.model.destination
 
-import org.opensearch.notifications.corespi.utils.validateMethod
+import org.opensearch.notifications.spi.utils.validateMethod
 
 /**
  * This class holds the contents of a custom webhook destination
