+++ notifications/core-spi/src/main/kotlin/org/opensearch/notifications/spi/model/destination/SlackDestination.kt
@@ -9,7 +9,7 @@
  *  GitHub history for details.
  */
 
-package org.opensearch.notifications.corespi.model.destination
+package org.opensearch.notifications.spi.model.destination
 
 /**
  * This class holds the contents of a Slack destination
