+++ notifications/core-spi/src/main/kotlin/org/opensearch/notifications/spi/model/destination/SmtpDestination.kt
@@ -25,11 +25,27 @@
  *
  */
 
-package org.opensearch.notifications.corespi.model.destination
+package org.opensearch.notifications.spi.model.destination
+
+import org.opensearch.common.Strings
+import org.opensearch.notifications.spi.utils.validateEmail
 
 /**
- * This class holds the contents of a Chime destination
+ * This class holds the contents of smtp destination
  */
-class ChimeDestination(
-    url: String,
-) : WebhookDestination(url, DestinationType.CHIME)
+class SmtpDestination(
+    val accountName: String,
+    val host: String,
+    val port: Int,
+    val method: String,
+    val fromAddress: String,
+    val recipient: String
+) : BaseDestination(DestinationType.SMTP) {
+
+    init {
+        require(!Strings.isNullOrEmpty(host)) { "Host name should be provided" }
+        require(port > 0) { "Port should be positive value" }
+        validateEmail(fromAddress)
+        validateEmail(recipient)
+    }
+}
