+++ notifications/core/src/main/kotlin/org/opensearch/notifications/core/client/DestinationSesClient.kt
@@ -26,9 +26,9 @@ import org.opensearch.notifications.core.credentials.SesClientFactory
 import org.opensearch.notifications.core.setting.PluginSettings
 import org.opensearch.notifications.core.utils.SecurityAccess
 import org.opensearch.notifications.core.utils.logger
-import org.opensearch.notifications.corespi.model.DestinationMessageResponse
-import org.opensearch.notifications.corespi.model.MessageContent
-import org.opensearch.notifications.corespi.model.destination.SesDestination
+import org.opensearch.notifications.spi.model.DestinationMessageResponse
+import org.opensearch.notifications.spi.model.MessageContent
+import org.opensearch.notifications.spi.model.destination.SesDestination
 import org.opensearch.rest.RestStatus
 import java.io.ByteArrayOutputStream
 import java.nio.ByteBuffer
