+++ notifications/core/src/main/kotlin/org/opensearch/notifications/core/client/DestinationSmtpClient.kt
@@ -16,10 +16,10 @@ import org.opensearch.common.settings.SecureString
 import org.opensearch.notifications.core.setting.PluginSettings
 import org.opensearch.notifications.core.utils.SecurityAccess
 import org.opensearch.notifications.core.utils.logger
-import org.opensearch.notifications.corespi.model.DestinationMessageResponse
-import org.opensearch.notifications.corespi.model.MessageContent
-import org.opensearch.notifications.corespi.model.SecureDestinationSettings
-import org.opensearch.notifications.corespi.model.destination.SmtpDestination
+import org.opensearch.notifications.spi.model.DestinationMessageResponse
+import org.opensearch.notifications.spi.model.MessageContent
+import org.opensearch.notifications.spi.model.SecureDestinationSettings
+import org.opensearch.notifications.spi.model.destination.SmtpDestination
 import org.opensearch.rest.RestStatus
 import java.util.Properties
 import javax.mail.Authenticator
