+++ notifications/core/src/main/kotlin/org/opensearch/notifications/core/client/EmailMessageValidator.kt
@@ -13,7 +13,7 @@ package org.opensearch.notifications.core.client
 
 import org.opensearch.notifications.core.setting.PluginSettings
 import org.opensearch.notifications.core.utils.logger
-import org.opensearch.notifications.corespi.model.MessageContent
+import org.opensearch.notifications.spi.model.MessageContent
 
 /**
  * This class handles the connections to the given Destination.
