+++ notifications/core/src/main/kotlin/org/opensearch/notifications/core/setting/PluginSettings.kt
@@ -22,7 +22,7 @@ import org.opensearch.common.settings.Settings
 import org.opensearch.notifications.core.NotificationCorePlugin.Companion.LOG_PREFIX
 import org.opensearch.notifications.core.NotificationCorePlugin.Companion.PLUGIN_NAME
 import org.opensearch.notifications.core.utils.logger
-import org.opensearch.notifications.corespi.model.SecureDestinationSettings
+import org.opensearch.notifications.spi.model.SecureDestinationSettings
 import java.io.IOException
 import java.nio.file.Path
