+++ notifications/core/src/test/kotlin/org/opensearch/notifications/core/SmtpDestinationTests.kt
@@ -22,11 +22,11 @@ import org.opensearch.common.settings.SecureString
 import org.opensearch.notifications.core.client.DestinationSmtpClient
 import org.opensearch.notifications.core.transport.DestinationTransportProvider
 import org.opensearch.notifications.core.transport.SmtpDestinationTransport
-import org.opensearch.notifications.corespi.model.DestinationMessageResponse
-import org.opensearch.notifications.corespi.model.MessageContent
-import org.opensearch.notifications.corespi.model.SecureDestinationSettings
-import org.opensearch.notifications.corespi.model.destination.DestinationType
-import org.opensearch.notifications.corespi.model.destination.SmtpDestination
+import org.opensearch.notifications.spi.model.DestinationMessageResponse
+import org.opensearch.notifications.spi.model.MessageContent
+import org.opensearch.notifications.spi.model.SecureDestinationSettings
+import org.opensearch.notifications.spi.model.destination.DestinationType
+import org.opensearch.notifications.spi.model.destination.SmtpDestination
 import org.opensearch.rest.RestStatus
 import javax.mail.MessagingException
