+++ notifications/core/src/test/kotlin/org/opensearch/notifications/core/integTest/SmtpEmailIT.kt
@@ -13,8 +13,8 @@ package org.opensearch.notifications.core.integTest
 
 import org.junit.After
 import org.opensearch.notifications.core.NotificationCoreImpl
-import org.opensearch.notifications.corespi.model.MessageContent
-import org.opensearch.notifications.corespi.model.destination.SmtpDestination
+import org.opensearch.notifications.spi.model.MessageContent
+import org.opensearch.notifications.spi.model.destination.SmtpDestination
 import org.opensearch.rest.RestStatus
 import org.opensearch.test.rest.OpenSearchRestTestCase
 import org.springframework.integration.test.mail.TestMailServer
