+++ notifications/notifications/src/main/kotlin/org/opensearch/notifications/send/SendMessageActionHelper.kt
@@ -45,15 +45,6 @@ import org.opensearch.commons.notifications.model.Webhook
 import org.opensearch.commons.utils.logger
 import org.opensearch.notifications.CoreProvider
 import org.opensearch.notifications.NotificationPlugin.Companion.LOG_PREFIX
-import org.opensearch.notifications.corespi.model.DestinationMessageResponse
-import org.opensearch.notifications.corespi.model.MessageContent
-import org.opensearch.notifications.corespi.model.destination.BaseDestination
-import org.opensearch.notifications.corespi.model.destination.ChimeDestination
-import org.opensearch.notifications.corespi.model.destination.CustomWebhookDestination
-import org.opensearch.notifications.corespi.model.destination.SesDestination
-import org.opensearch.notifications.corespi.model.destination.SlackDestination
-import org.opensearch.notifications.corespi.model.destination.SmtpDestination
-import org.opensearch.notifications.corespi.model.destination.SnsDestination
 import org.opensearch.notifications.index.ConfigOperations
 import org.opensearch.notifications.index.EventOperations
 import org.opensearch.notifications.metrics.Metrics
@@ -61,6 +52,15 @@ import org.opensearch.notifications.model.DocMetadata
 import org.opensearch.notifications.model.NotificationConfigDocInfo
 import org.opensearch.notifications.model.NotificationEventDoc
 import org.opensearch.notifications.security.UserAccess
+import org.opensearch.notifications.spi.model.DestinationMessageResponse
+import org.opensearch.notifications.spi.model.MessageContent
+import org.opensearch.notifications.spi.model.destination.BaseDestination
+import org.opensearch.notifications.spi.model.destination.ChimeDestination
+import org.opensearch.notifications.spi.model.destination.CustomWebhookDestination
+import org.opensearch.notifications.spi.model.destination.SesDestination
+import org.opensearch.notifications.spi.model.destination.SlackDestination
+import org.opensearch.notifications.spi.model.destination.SmtpDestination
+import org.opensearch.notifications.spi.model.destination.SnsDestination
 import org.opensearch.rest.RestStatus
 import java.io.ByteArrayOutputStream
 import java.time.Instant
