+++ notifications/notifications/src/main/resources/META-INF/services/org.opensearch.notifications.spi.NotificationCoreExtension
